Add API:RangeSearch,search,/users/{userId}

+++ src/main/java/cn/edu/bjut/librarymanagementsystem/dto/BookQueryByIdRequest.java
@@ -1,6 +1,6 @@
 package cn.edu.bjut.librarymanagementsystem.dto;
 
 public record BookQueryByIdRequest(
-    Long startId,
-    Long endId
+    int startId,
+    int endId
 ) {}
